+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from "./native-modules/advertise";
 export * from "./native-modules/connection";
 export * from "./native-modules/discovery";
+export * from "./native-modules/file-transfer";
 export * from "./native-modules/permission";
 export * from "./types/nearby-connections.types";
 export type { Unsubscribe } from "./utilities/create-event-handler";